pyproject.toml
[project]
name = "chatbot-ai-assistant-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.121.1",
    "greenlet>=3.2.4",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "langchain>=1.0.8",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=3.1.0",
    "langchain-postgres>=0.0.16",
    "langgraph>=1.0.3",
    "pgvector>=0.3.6",
    "psycopg[binary]>=3.2.13",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
    "sqlalchemy>=2.0.44",
    "tavily-python>=0.7.13",
    "langchain-tavily>=0.0.2",
    "uvicorn>=0.38.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.2",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
]
